On déplace les composants dans un dossier DPE, on sépare DPEMap et DPEAddress
