+++ components/dpe/DPEFacture.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import data from '@/components/dpe/DPE.yaml'
 import { formatNumberWithSpaces } from '../utils'
 import { formatNumber } from '../RevenuInput'
-import DPEQuickSwitch from '../dpe/DPEQuickSwitch'
+import DPEQuickSwitch from './DPEQuickSwitch'
 import TargetDPETabs from '../mpra/TargetDPETabs'
 import rules from '@/app/règles/rules'
 import Select from '../Select'
